changed latitude and longitude properties from long to double on he Branch DAO object

+++ android-app/src/com/exchange/dao/Branch.java
@@ -9,10 +9,10 @@ public class Branch {
     private long id;
     /** Not-null value. */
     private String name;
-    private Long latitude;
-    private Long longitude;
+    private Double latitude;
+    private Double longitude;
     private String address;
-    private Long bankId;
+    private long bankId;
 
     public Branch() {
     }
@@ -21,7 +21,7 @@ public class Branch {
         this.id = id;
     }
 
-    public Branch(long id, String name, Long latitude, Long longitude, String address, Long bankId) {
+    public Branch(long id, String name, Double latitude, Double longitude, String address, long bankId) {
         this.id = id;
         this.name = name;
         this.latitude = latitude;
@@ -48,19 +48,19 @@ public class Branch {
         this.name = name;
     }
 
-    public Long getLatitude() {
+    public Double getLatitude() {
         return latitude;
     }
 
-    public void setLatitude(Long latitude) {
+    public void setLatitude(Double latitude) {
         this.latitude = latitude;
     }
 
-    public Long getLongitude() {
+    public Double getLongitude() {
         return longitude;
     }
 
-    public void setLongitude(Long longitude) {
+    public void setLongitude(Double longitude) {
         this.longitude = longitude;
     }
 
@@ -72,11 +72,11 @@ public class Branch {
         this.address = address;
     }
 
-    public Long getBankId() {
+    public long getBankId() {
         return bankId;
     }
 
-    public void setBankId(Long bankId) {
+    public void setBankId(long bankId) {
         this.bankId = bankId;
     }
